pyproject.toml
[tool.poetry]
name = "rent-games"
version = "0.1.0"
description = ""
authors = ["dmkisel <i@dkisel.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
telebot = "^0.0.5"
celery = "^5.4.0"
asyncio = "^3.4.3"
streamlit = "^1.36.0"
pg8000 = "^1.31.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
yookassa = "^3.3.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
